refactor: make pdf document files reusable
refactor: add missing modifications to downloadPdf
