+++ src/components/Report/PdfReports/ConstructionProgram/index.ts
@@ -1,1 +1,1 @@
-export { default } from './ConstructionProgram';
+export { default } from '../ReportContainer';
